+++ src/lxc/bdev/lxcbtrfs.c
@@ -37,6 +37,7 @@
 #include "bdev.h"
 #include "log.h"
 #include "lxcbtrfs.h"
+#include "lxcrsync.h"
 #include "utils.h"
 
 lxc_log_define(btrfs, lxc);
